+++ EFFoundation/Classes/CGSize+.swift
@@ -1,6 +1,6 @@
 //
 //  CGSize+.swift
-//  DeviceKit
+//  EFFoundation
 //
 //  Created by EyreFree on 2019/8/14.
 //