+++ EFFoundation/Classes/NSMutableAttributedString+.swift
@@ -1,6 +1,6 @@
 //
 //  NSMutableAttributedString+.swift
-//  DeviceKit
+//  EFFoundation
 //
 //  Created by EyreFree on 2019/8/14.
 //