+++ EFFoundation/Classes/NSString+.swift
@@ -1,6 +1,6 @@
 //
 //  NSString+.swift
-//  DeviceKit
+//  EFFoundation
 //
 //  Created by EyreFree on 2019/8/14.
 //