+++ EFFoundation/Classes/String+.swift
@@ -1,6 +1,6 @@
 //
 //  String+.swift
-//  DeviceKit
+//  EFFoundation
 //
 //  Created by EyreFree on 2019/8/14.
 //