+++ EFFoundation/Classes/UIColor+.swift
@@ -1,6 +1,6 @@
 //
 //  UIColor+.swift
-//  DeviceKit
+//  EFFoundation
 //
 //  Created by EyreFree on 2019/8/15.
 //