+++ EFFoundation/Classes/UIImage+.swift
@@ -1,6 +1,6 @@
 //
 //  UIImage+.swift
-//  DeviceKit
+//  EFFoundation
 //
 //  Created by EyreFree on 2019/8/15.
 //