+++ CHANGELOG.md
@@ -2,9 +2,18 @@
 
 -----
 
+## [0.4.0](https://github.com/EFPrefix/EFFoundation/releases/tag/0.4.0) (2019-08-20)
+
+- Add extension for `BinaryFloatingPoint`;
+- Add extension for `UIGestureRecognizer`.
+
+---
+
 ## [0.3.0](https://github.com/EFPrefix/EFFoundation/releases/tag/0.3.0) (2019-08-16)
 
 - Add dependency on SwiftyAttributes.
+
+---
 
 ## [0.1.0](https://github.com/EFPrefix/EFFoundation/releases/tag/0.1.0) (2019-08-15)